stt/pyproject.toml
[project]
name = "stt"
version = "0.1.0"
description = """
STT Service is an asynchronous microservice for speech-to-text processing.
It leverages OpenAI Whisper for accurate transcription, Pydub for audio
manipulation, and FastStream with RabbitMQ for message exchange. Dishka
provides dependency injection and modular design, while Pydantic enforces
strict DTO validation. Uvicorn serves as the runtime for exposing APIs and
integrations. This service acts as the audio input layer, converting voice
commands into structured text for downstream services such as NLU and
Command.
"""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dishka>=1.7.2",
    "faststream[rabbit]>=0.6.4",
    "openai-whisper>=20250625",
    "pydub>=0.25.1",
    "redis>=7.1.0",
    "uvicorn>=0.39.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.10",
    "types-redis>=4.6.0.20241004",
]

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = ["migrations/*", "tests/*"]
fix = true

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E203"]
preview = true

[tool.ruff.lint.isort]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
known-first-party = ["stt"]
combine-as-imports = true
force-sort-within-sections = true
lines-after-imports = 2

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = 88

[tool.mypy]
ignore_missing_imports = true
explicit_package_bases = true 
namespace_packages = true


